# docs/swagger.yaml
basePath: /
definitions:
...
  models.ContactMessage:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      message:
        type: string
      name:
        type: string
      resolved_by:
        type: string
      status:
        description: '"new", "in_progress", "resolved"'
        type: string
      subject:
        type: string
      updated_at:
        type: string
    required:
    - email
    - message
    - name
    type: object
...
  models.LocalizedContent:
    type: object
  models.LocalizedField:
    properties:
      content:
        type: string
      slug:
        type: string
      title:
        type: string
    type: object
...
  models.Media:
    properties:
      file_name:
        type: string
      file_path:
        type: string
      file_size:
        example: 102400
        type: integer
      file_type:
        type: string
      id:
        type: string
      uploaded_at:
        type: integer
      uploaded_by:
        example: admin
        type: string
    type: object
...
info:
  contact:
    email: support@example.com
    name: API Support
  description: This is the API documentation for the admin panel project.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Admin Panel API
  version: "1.0"
paths:
...
  /contacts:
    get:
      description: Retrieve all contact messages sent by users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ContactMessage'
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all contact messages
      tags:
      - Contacts
    post:
      consumes:
      - application/json
      description: Add a new contact message sent by a user
      parameters:
      - description: Contact message details
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.ContactMessage'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ContactMessage'
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Create a new contact message
      tags:
      - Contacts
  /contacts/{id}:
    delete:
      description: Remove a contact message by its unique identifier
      parameters:
      - description: Contact ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid contact ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Contact message not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Delete a contact message
      tags:
      - Contacts
    get:
      description: Retrieve a single contact message by its unique identifier
      parameters:
      - description: Contact Message ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ContactMessage'
        "400":
          description: Invalid contact ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Contact message not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get a contact message by ID
      tags:
      - Contacts
...
  /localized-content:
    post:
      consumes:
      - application/json
      description: Add new localized content with translations
      parameters:
      - description: Localized content details
        in: body
        name: content
        required: true
        schema:
          $ref: '#/definitions/models.LocalizedContent'
      produces:
      - application/json
      responses:
        "200":
          description: Content created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request payload
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create content
          schema:
            additionalProperties: true
            type: object
      summary: Create localized content
      tags:
      - Localized Content
  /localized-content/{id}:
    get:
      description: Retrieve localized content in a specific language
      parameters:
      - description: Localized Content ID
        in: path
        name: id
        required: true
        type: string
      - description: Language code (e.g., 'en', 'tr')
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Localized content details
          schema:
            $ref: '#/definitions/models.LocalizedField'
        "400":
          description: Invalid content ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Content not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch content
          schema:
            additionalProperties: true
            type: object
      summary: Get localized content by ID
      tags:
      - Localized Content
...
  /media:
    get:
      description: Retrieve all media files in the library
      produces:
      - application/json
      responses:
        "200":
          description: List of media files
          schema:
            items:
              $ref: '#/definitions/models.Media'
            type: array
        "500":
          description: Failed to retrieve media files
          schema:
            additionalProperties: true
            type: object
      summary: Get all media files
      tags:
      - Media
  /media/{id}:
    delete:
      description: Remove a specific media file by its ID
      parameters:
      - description: Media file ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid media ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Media file not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete media file
          schema:
            additionalProperties: true
            type: object
      summary: Delete media file
      tags:
      - Media
    get:
      description: Retrieve details of a specific media file by its ID
      parameters:
      - description: Media file ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Media file details
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: Invalid media ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Media file not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to retrieve media details
          schema:
            additionalProperties: true
            type: object
      summary: Get media details
      tags:
      - Media
  /media/filter:
    get:
      description: Retrieve media files filtered by file name, type, or upload date
      parameters:
      - description: Filter by file name
        in: query
        name: file_name
        type: string
      - description: Filter by file type
        in: query
        name: file_type
        type: string
      - description: Start date for upload filter (YYYY-MM-DD)
        in: query
        name: start_date
        type: string
      - description: End date for upload filter (YYYY-MM-DD)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Filtered list of media files
          schema:
            items:
              $ref: '#/definitions/models.Media'
            type: array
        "500":
          description: Failed to retrieve filtered media files
          schema:
            additionalProperties: true
            type: object
      summary: Filter media files
      tags:
      - Media
  /media/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to the media library
      parameters:
      - description: Media file to upload
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: File uploaded successfully
          schema:
            $ref: '#/definitions/models.Media'
        "400":
          description: No file uploaded or invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to save file or record
          schema:
            additionalProperties: true
            type: object
      summary: Upload a new media file
      tags:
      - Media
...
security:
- BearerAuth: []
swagger: "2.0"